Add Breed and Pet controllers, fix small bugs, change database structure (#5)
* Add Breed and Pet controllers, fix small bugs, change database structure

* Fix role hierarchy to not fetch roles from db every time

* Fix logs warnings

* Add swagger annotations to enhance documentation

+++ src/main/java/com/peti/backend/model/City.java
@@ -18,7 +18,7 @@ public class City {
   @Id
   @EqualsAndHashCode.Include
   @Column(name = "city_id", nullable = false)
-  private int cityId;
+  private Long cityId;
 
   @Basic
   @Column(name = "longitude", precision = 8)
